Conflicts:
	.idea/libraries/actionbarsherlock_4_4_0.xml
	.idea/misc.xml
	GPSLogger/GPSLogger.iml
	GPSLogger/src/main/java/com/mendhak/gpslogger/loggers/AbstractLiveLogger.java
	GPSLogger/src/main/java/com/mendhak/gpslogger/loggers/LiveTrack24FileLogger.java
	GPSLogger/src/main/java/com/mendhak/gpslogger/loggers/utils/LocationBuffer.java
	build.gradle
	gradle/wrapper/gradle-wrapper.properties

+++ GPSLogger/src/main/java/com/mendhak/gpslogger/loggers/SkyLinesLogger.java
@@ -28,6 +28,7 @@ package com.mendhak.gpslogger.loggers;
 import android.location.Location;
 import android.os.SystemClock;
 
+import com.mendhak.gpslogger.common.IActionListener;
 import com.mendhak.gpslogger.common.Utilities;
 import com.mendhak.gpslogger.loggers.utils.LocationBuffer;
 
@@ -203,7 +204,7 @@ public class SkyLinesLogger extends AbstractLiveLogger
                 b.lat, b.lon,
                 b.altitude,
                 b.bearing,
-                b.speed);
+                (int)b.speed);
         return true;
     }
